--- packages/DropdownMenu/package.json
@@ -1,46 +1,0 @@
-{
-  "name": "@paprika/dropdown-menu",
-  "version": "2.0.47-alpha.4",
-  "description": "DropdownMenu component displays a trigger button, which when clicked displays a list of items in a dropdown format. These items can be raw content, Links or actions such as Delete, which will prompt a confirmation panel to be displayed.",
-  "author": "@paprika",
-  "license": "MIT",
-  "main": "lib/index.js",
-  "types": "lib/index.d.ts",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/acl-services/paprika.git",
-    "directory": "packages/DropdownMenu"
-  },
-  "publishConfig": {
-    "access": "public"
-  },
-  "dependencies": {
-    "@babel/runtime-corejs2": "^7.3.1",
-    "@paprika/button": "^0.3.15-alpha.1",
-    "@paprika/confirmation": "^0.3.44-alpha.2",
-    "@paprika/constants": "^0.1.3",
-    "@paprika/helpers": "^0.2.13",
-    "@paprika/icon": "^0.3.11-alpha.0",
-    "@paprika/popover": "^0.3.36-alpha.0",
-    "@paprika/raw-button": "^0.2.20",
-    "@paprika/stylers": "^0.2.11",
-    "@paprika/tokens": "^0.1.17",
-    "prop-types": "^15.7.2",
-    "uuid": "^3.3.2"
-  },
-  "peerDependencies": {
-    "@paprika/l10n": "^0.2.5",
-    "react": "^16.8.4",
-    "react-dom": "^16.8.4",
-    "styled-components": "^4.2.0"
-  },
-  "paprikaDocs": {
-    "subComponents": [
-      "Divider",
-      "Item",
-      "LinkItem",
-      "Trigger"
-    ]
-  },
-  "gitHead": "494ce26cb0a6c24a8f6abf2b154d97ddd4d7e343"
-}
