--- packages/DropdownMenu/src/components/Content/Content.js
@@ -1,7 +1,0 @@
-import React from "react";
-
-export default function Content() {
-  return <React.Fragment />;
-}
-
-Content.displayName = "DropdownMenu.Content";
