--- packages/DropdownMenu/src/components/Divider/Divider.js
@@ -1,8 +1,0 @@
-import React from "react";
-import DividerStyles from "./Divider.styles";
-
-export default function Divider() {
-  return <span css={DividerStyles} />;
-}
-
-Divider.displayName = "DropdownMenu.Divider";
