--- packages/DropdownMenu/stories/DropdownMenu.stories.styles.js
@@ -1,8 +1,0 @@
-import styled from "styled-components";
-import { Story } from "storybook/assets/styles/common.styles";
-
-// Common DropdownMenu story styles
-
-export const DropdownMenuStory = styled(Story)`
-  max-width: 500px;
-`;
