--- packages/DropdownMenu/stories/examples/DropdownMenuLongestExample.js
@@ -1,26 +1,0 @@
-import React from "react";
-import DropdownMenu from "../../src";
-
-const DropdownMenuExample = () => {
-  return (
-    <DropdownMenu align="bottom">
-      <DropdownMenu.Trigger>Trigger</DropdownMenu.Trigger>
-      <DropdownMenu.Item onClick={() => {}}>
-        SupercalifragilisticexpialidociousSupercalifragilisticexpialidociousSupercalifragilisticexpialidocious
-      </DropdownMenu.Item>
-      <DropdownMenu.Item onClick={() => {}}>Pseudopseudohypoparathyroidism</DropdownMenu.Item>
-      <DropdownMenu.Item onClick={() => {}}>Floccinaucinihilipilification</DropdownMenu.Item>
-      <DropdownMenu.Item onClick={() => {}}>Antidisestablishmentarianism</DropdownMenu.Item>
-      <DropdownMenu.Item onClick={() => {}}>Supercalifragilisticexpialidocious</DropdownMenu.Item>
-      <DropdownMenu.Item onClick={() => {}}>Pseudopseudohypoparathyroidism</DropdownMenu.Item>
-      <DropdownMenu.Item onClick={() => {}}>Floccinaucinihilipilification</DropdownMenu.Item>
-      <DropdownMenu.Item onClick={() => {}}>Antidisestablishmentarianism</DropdownMenu.Item>
-      <DropdownMenu.Item onClick={() => {}}>Supercalifragilisticexpialidocious</DropdownMenu.Item>
-      <DropdownMenu.Item onClick={() => {}}>Pseudopseudohypoparathyroidism</DropdownMenu.Item>
-      <DropdownMenu.Item onClick={() => {}}>Floccinaucinihilipilification</DropdownMenu.Item>
-      <DropdownMenu.Item onClick={() => {}}>Antidisestablishmentarianism</DropdownMenu.Item>
-    </DropdownMenu>
-  );
-};
-
-export default DropdownMenuExample;
