+++ packages/OverflowMenu/src/components/Content/Content.js
@@ -1,0 +1,7 @@
+import React from "react";
+
+export default function Content() {
+  return <React.Fragment />;
+}
+
+Content.displayName = "OverflowMenu.Content";
