+++ packages/OverflowMenu/src/components/Divider/Divider.js
@@ -1,0 +1,8 @@
+import React from "react";
+import DividerStyles from "./Divider.styles";
+
+export default function Divider() {
+  return <span css={DividerStyles} />;
+}
+
+Divider.displayName = "OverflowMenu.Divider";
