+++ packages/OverflowMenu/stories/OverflowMenu.stories.styles.js
@@ -1,0 +1,8 @@
+import styled from "styled-components";
+import { Story } from "storybook/assets/styles/common.styles";
+
+// Common OverflowMenu story styles
+
+export const OverflowMenuStory = styled(Story)`
+  max-width: 500px;
+`;
