+++ packages/OverflowMenu/stories/examples/OverflowMenuLongestExample.js
@@ -1,0 +1,26 @@
+import React from "react";
+import OverflowMenu from "../../src";
+
+const OverflowMenuExample = () => {
+  return (
+    <OverflowMenu align="bottom">
+      <OverflowMenu.Trigger>Trigger</OverflowMenu.Trigger>
+      <OverflowMenu.Item onClick={() => {}}>
+        SupercalifragilisticexpialidociousSupercalifragilisticexpialidociousSupercalifragilisticexpialidocious
+      </OverflowMenu.Item>
+      <OverflowMenu.Item onClick={() => {}}>Pseudopseudohypoparathyroidism</OverflowMenu.Item>
+      <OverflowMenu.Item onClick={() => {}}>Floccinaucinihilipilification</OverflowMenu.Item>
+      <OverflowMenu.Item onClick={() => {}}>Antidisestablishmentarianism</OverflowMenu.Item>
+      <OverflowMenu.Item onClick={() => {}}>Supercalifragilisticexpialidocious</OverflowMenu.Item>
+      <OverflowMenu.Item onClick={() => {}}>Pseudopseudohypoparathyroidism</OverflowMenu.Item>
+      <OverflowMenu.Item onClick={() => {}}>Floccinaucinihilipilification</OverflowMenu.Item>
+      <OverflowMenu.Item onClick={() => {}}>Antidisestablishmentarianism</OverflowMenu.Item>
+      <OverflowMenu.Item onClick={() => {}}>Supercalifragilisticexpialidocious</OverflowMenu.Item>
+      <OverflowMenu.Item onClick={() => {}}>Pseudopseudohypoparathyroidism</OverflowMenu.Item>
+      <OverflowMenu.Item onClick={() => {}}>Floccinaucinihilipilification</OverflowMenu.Item>
+      <OverflowMenu.Item onClick={() => {}}>Antidisestablishmentarianism</OverflowMenu.Item>
+    </OverflowMenu>
+  );
+};
+
+export default OverflowMenuExample;
